+++ src/continuiti/operators/fourierlayer.py
@@ -111,8 +111,6 @@ class FourierLayer(Operator):
 
     """
 
-    TENSOR_INDICES = "defghijklmnopqrstuvxyz"
-
     def __init__(
         self,
         shapes: OperatorShapes,
@@ -197,7 +195,7 @@ class FourierLayer(Operator):
         )
 
         # perform kernel integral operation in Fourier space
-        out_fft = self._contract_with_kernel(u_fft, dim=fft_dimensions)
+        out_fft = self._contract_with_kernel(u_fft)
 
         # the output shape is determined by y.shape[2:] (num_evaluations...) and not x.shape[2:] (num_sensors...)
         target_shape = list(y.shape[2:])
@@ -229,34 +227,21 @@ class FourierLayer(Operator):
         return out
 
     def _contract_with_kernel(
-        self, fft_values: torch.Tensor, dim: Tuple[int]
+        self,
+        fft_values: torch.Tensor,
     ) -> torch.Tensor:
         """Contract kernel with input values.
 
         Args:
             fft_values: Tensor with fft values.
-            dim: List of fft dimensions.
 
         Returns:
             Output tensor with shape = (batch_size, ..., v.dim)
         """
 
-        num_fft_dimensions = len(dim)
-
-        assert (
-            len(self.TENSOR_INDICES) > num_fft_dimensions
-        ), f"Too many dimensions. The current limit for the number of dimensions is {len(self.TENSOR_INDICES)}."
-
-        # contraction equation for torch.einsum method
-        # d: v-dim, s: u-dim, b: batch-dim
-        frequency_indices = "".join(self.TENSOR_INDICES[: int(num_fft_dimensions)])
-        contraction_equation = "{}ac,b{}c->b{}a".format(
-            frequency_indices, frequency_indices, frequency_indices
-        )
-
         # perform kernel operation in Fourier space
         kernel = torch.view_as_complex(self.kernel)
-        out_fft = torch.einsum(contraction_equation, kernel, fft_values)
+        out_fft = torch.einsum("...ac,b...c->b...a", kernel, fft_values)
 
         return out_fft